requirements.txt
pyannote.audio
openai-whisper
python-dotenv
